# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Xuthority API",
    "version": "1.0.0",
    "description": "Automatically generated API documentation.",
    "contact": {
      "name": "Xuthority",
      "url": "https://xuthority.com",
      "email": "support@xuthority.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://xuthority.com/terms",
    "xuthority": "https://xuthority.com",
    "xuthorityDev": "https://xuthority-dev.com",
    "xuthorityDevApi": "http://localhost:8081/api/v1",
    "xuthorityDevApiDocs": "http://localhost:8081/api-docs",
    "xuthorityDevApiDocsSwagger": "http://localhost:8081/swagger.json",
    "xuthorityDevApiDocsSwaggerUi": "http://localhost:8081/api-docs",
    "xuthorityDevApiDocsSwaggerUiSwagger": "http://localhost:8081/swagger.json"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/disputes": {
      "post": {
        "summary": "Create a new dispute on a product review",
        "tags": [
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reviewId",
                  "reason",
                  "description"
                ],
                "properties": {
                  "reviewId": {
                    "type": "string",
                    "description": "ID of the review to dispute"
                  },
                  "reason": {
                    "type": "string",
                    "enum": [
                      "false-or-misleading-information",
                      "spam-or-fake-review",
                      "inappropriate-content",
                      "conflict-of-interest",
                      "other"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dispute created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not product owner"
          }
        }
      },
      "get": {
        "summary": "Get vendor's disputes with pagination",
        "tags": [
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "resolved"
              ]
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "status"
              ],
              "default": "createdAt"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Disputes retrieved successfully"
          }
        }
      }
    },
    "/api/v1/disputes/all": {
      "get": {
        "summary": "Get all disputes (Admin only)",
        "tags": [
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "resolved"
              ]
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "status"
              ],
              "default": "createdAt"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All disputes retrieved successfully"
          }
        }
      }
    },
    "/api/v1/disputes/{id}": {
      "get": {
        "summary": "Get dispute by ID",
        "tags": [
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dispute retrieved successfully"
          },
          "404": {
            "description": "Dispute not found"
          }
        }
      },
      "put": {
        "summary": "Update dispute",
        "tags": [
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "enum": [
                      "false-or-misleading-information",
                      "spam-or-fake-review",
                      "inappropriate-content",
                      "conflict-of-interest",
                      "other"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "resolved"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dispute updated successfully"
          },
          "404": {
            "description": "Dispute not found"
          }
        }
      },
      "delete": {
        "summary": "Delete dispute",
        "tags": [
          "Disputes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dispute deleted successfully"
          },
          "404": {
            "description": "Dispute not found"
          }
        }
      }
    },
...
  },
...
}